requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
decorator==5.1.1
Django==5.0.4
django-celery-beat==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
executing==2.0.1
ipython==8.23.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
redis==5.0.3
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
